refactor: extract cut handler and image pick hook from RichMarkdownEditor
