+++ src/renderer/src/components/editor/RichMarkdownEditor.tsx
@@ -1,12 +1,11 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { EditorContent, useEditor } from '@tiptap/react'
 import type { Editor } from '@tiptap/react'
-import { toast } from 'sonner'
 import { RichMarkdownSlashMenu } from './RichMarkdownSlashMenu'
 import { useAppStore } from '@/store'
 import { RichMarkdownToolbar } from './RichMarkdownToolbar'
-import { extractIpcErrorMessage, getImageCopyDestination } from './rich-markdown-image-utils'
 import { encodeRawMarkdownHtmlForRichEditor } from './raw-markdown-html'
+import { useLocalImagePick } from './useLocalImagePick'
 import { createRichMarkdownExtensions } from './rich-markdown-extensions'
 import { slashCommands, syncSlashMenu } from './rich-markdown-commands'
 import type { SlashCommand, SlashMenuState } from './rich-markdown-commands'
@@ -21,11 +20,9 @@ import { useLinkBubble } from './useLinkBubble'
 import { useEditorScrollRestore } from './useEditorScrollRestore'
 import { registerPendingEditorFlush } from './editor-pending-flush'
 import { createRichMarkdownKeyHandler } from './rich-markdown-key-handler'
-import { DOMSerializer } from '@tiptap/pm/model'
-import { TextSelection } from '@tiptap/pm/state'
 import { normalizeSoftBreaks } from './rich-markdown-normalize'
 import { autoFocusRichEditor } from './rich-markdown-auto-focus'
-import { cutVisualLine, getVisualLineRange } from './rich-markdown-visual-line'
+import { handleRichMarkdownCut } from './rich-markdown-cut-handler'
 
 type RichMarkdownEditorProps = {
   fileId: string
@@ -119,110 +116,8 @@ export default function RichMarkdownEditor({
       attributes: {
         class: 'rich-markdown-editor'
       },
-      // Why: Electron's app menu `{ role: 'cut' }` binds Cmd/Ctrl+X at the
-      // main-process level, so the keystroke never reaches handleKeyDown.
-      // Instead, the menu dispatches a native cut command which fires this
-      // DOM event. For empty selections we cut the current block (like VS
-      // Code and Notion); for non-empty selections we defer to ProseMirror's
-      // built-in clipboard serializer.
       handleDOMEvents: {
-        cut: (view, event) => {
-          const { selection } = view.state
-          if (!selection.empty) {
-            return false
-          }
-
-          const { $from } = selection
-
-          // Why: a GapCursor before a top-level leaf node (e.g. horizontal rule
-          // as the first child of the doc) resolves to depth 0. Attempting to cut
-          // at depth 0 would call $from.before(0) on the doc node, which throws
-          // RangeError("There is no position before the top-level node").  Bail
-          // out and let ProseMirror's default handler deal with it.
-          if ($from.depth < 1) {
-            return false
-          }
-
-          // Walk up from the textblock to find the best node to cut. For list
-          // items and task items, cut the whole item rather than just its inner
-          // paragraph. Stop at table cells to avoid breaking table structure.
-          let cutDepth = $from.depth
-          for (let d = $from.depth - 1; d >= 1; d--) {
-            const name = $from.node(d).type.name
-            if (name === 'listItem' || name === 'taskItem') {
-              cutDepth = d
-              break
-            }
-            if (name === 'tableCell' || name === 'tableHeader') {
-              break
-            }
-          }
-
-          const cutNode = $from.node(cutDepth)
-          const text = cutNode.textContent
-
-          // Why: for paragraphs that word-wrap across multiple visual lines, cut
-          // only the visual line the cursor is on rather than the entire paragraph.
-          // This matches the user expectation of per-line cutting (like VS Code)
-          // without destroying the rest of the paragraph's content.
-          if (cutNode.type.name === 'paragraph' && text) {
-            const paraStart = $from.start(cutDepth)
-            const paraEnd = $from.end(cutDepth)
-            const lineRange = getVisualLineRange(view, selection.from, paraStart, paraEnd)
-            if (lineRange) {
-              return cutVisualLine(view, event, lineRange)
-            }
-            // Falls through to block-level cut for single-line paragraphs.
-          }
-
-          if (!text) {
-            // Still delete the empty block, matching VS Code behavior
-            event.preventDefault()
-            const from = $from.before(cutDepth)
-            const to = $from.after(cutDepth)
-            let tr = view.state.tr.delete(from, to)
-            // Why: after deleting the last block the old `from` offset may exceed
-            // the new document length, so we clamp and use TextSelection.near() to
-            // land on the closest valid cursor position.
-            const clampedPos = Math.max(0, Math.min(from, tr.doc.content.size))
-            const resolvedPos = tr.doc.resolve(clampedPos)
-            tr = tr.setSelection(TextSelection.near(resolvedPos))
-            view.dispatch(tr)
-            return true
-          }
-
-          const clipboardEvent = event as ClipboardEvent
-          // Why: if clipboardData is null (e.g. synthetic events), we must not
-          // preventDefault and then delete -- that would lose content without
-          // placing it on the clipboard. Fall back to browser default instead.
-          if (!clipboardEvent.clipboardData) {
-            return false
-          }
-          event.preventDefault()
-
-          // Why: writing both text/html and text/plain preserves inline formatting
-          // (bold, italic, links) on round-trip cut-then-paste, while still giving
-          // a plain-text fallback for external targets.
-          const serializer = DOMSerializer.fromSchema(view.state.schema)
-          const fragment = serializer.serializeFragment(cutNode.content)
-          const div = document.createElement('div')
-          div.appendChild(fragment)
-          clipboardEvent.clipboardData.setData('text/html', div.innerHTML)
-          clipboardEvent.clipboardData.setData('text/plain', text)
-
-          const from = $from.before(cutDepth)
-          const to = $from.after(cutDepth)
-          let tr = view.state.tr.delete(from, to)
-          // Why: after deleting the last block the old `from` offset may exceed
-          // the new document length, so we clamp and use TextSelection.near() to
-          // land on the closest valid cursor position.
-          const clampedPos = Math.max(0, Math.min(from, tr.doc.content.size))
-          const resolvedPos = tr.doc.resolve(clampedPos)
-          tr = tr.setSelection(TextSelection.near(resolvedPos))
-          view.dispatch(tr)
-
-          return true
-        }
+        cut: handleRichMarkdownCut
       },
       handleKeyDown: createRichMarkdownKeyHandler({
         isMac,
@@ -281,26 +176,19 @@ export default function RichMarkdownEditor({
     onUpdate: ({ editor: nextEditor }) => {
       syncSlashMenu(nextEditor, rootRef.current, setSlashMenu)
 
-      // Why: normalizeSoftBreaks in onCreate dispatches a transaction that
-      // triggers this callback before the editor is ready for user interaction.
-      // Treating that structural housekeeping step as a user edit would
-      // immediately mark the file dirty and prompt a spurious save dialog on
-      // close. Bail out until initialization is complete.
+      // Why: bail out during normalizeSoftBreaks's onCreate transaction so the
+      // structural housekeeping doesn't mark the file dirty before the user
+      // has typed anything.
       if (isInitializingRef.current) {
         return
       }
 
-      // Why: full markdown serialization is debounced for typing performance,
-      // but close-confirmation and beforeunload checks still need to know
-      // immediately that the document changed. Optimistically mark the tab
-      // dirty here, then let the debounced content sync compute the exact
-      // saved-vs-draft comparison a moment later.
+      // Why: optimistically mark dirty for close-confirmation before the
+      // debounced content sync computes the exact saved-vs-draft comparison.
       onDirtyStateHintRef.current(true)
 
-      // Why: getMarkdown() serializes the entire ProseMirror document tree on
-      // every keystroke, which is the dominant typing-speed bottleneck for large
-      // files. Debouncing it to 300ms keeps the draft store and autosave pipeline
-      // fed without blocking the input path.
+      // Why: getMarkdown() is the typing-speed bottleneck for large files;
+      // debouncing to 300ms keeps drafts current without blocking input.
       if (serializeTimerRef.current !== null) {
         window.clearTimeout(serializeTimerRef.current)
       }
@@ -338,13 +226,10 @@ export default function RichMarkdownEditor({
     editorRef.current = editor ?? null
   }, [editor])
 
-  // Why: when the component unmounts (tab switch, mode change), flush any
-  // pending serialization so the autosave controller's draft store has the
-  // latest content and the scroll-position cache captures the right state.
-  // This must run before useEditor's cleanup destroys the editor instance,
-  // so we use useLayoutEffect (synchronous cleanup) instead of useEffect.
-  // React runs layout-effect cleanups before effect cleanups, guaranteeing
-  // the editor is still alive when we serialize.
+  // Why: use useLayoutEffect (synchronous cleanup) so the pending serialization
+  // flush runs before useEditor's cleanup destroys the editor instance on tab
+  // switch or mode change. React runs layout-effect cleanups before effect
+  // cleanups, guaranteeing the editor is still alive when we serialize.
   React.useLayoutEffect(() => {
     return flushPendingSerialization
   }, [flushPendingSerialization])
@@ -363,38 +248,7 @@ export default function RichMarkdownEditor({
     }
   }, [editor, filePath])
 
-  const handleLocalImagePick = useCallback(async () => {
-    if (!editor) {
-      return
-    }
-    // Why: the native file picker steals focus from the editor, which can cause
-    // ProseMirror to lose track of its selection. We snapshot the cursor position
-    // before the async dialog so we can insert the image exactly where the user
-    // intended, not at whatever position focus() falls back to afterward.
-    const insertPos = editor.state.selection.from
-    try {
-      const srcPath = await window.api.shell.pickImage()
-      if (!srcPath) {
-        return
-      }
-      // Why: copy the image next to the markdown file and insert a relative path
-      // so the markdown stays portable and doesn't bloat with base64 data.
-      const { imageName, destPath } = await getImageCopyDestination(filePath, srcPath)
-      if (srcPath !== destPath) {
-        await window.api.shell.copyFile({ srcPath, destPath })
-      }
-      // Why: insertContentAt places the image at the exact saved position
-      // regardless of where focus lands after the native file dialog closes,
-      // whereas setTextSelection can be overridden by ProseMirror's focus logic.
-      editor
-        .chain()
-        .focus()
-        .insertContentAt(insertPos, { type: 'image', attrs: { src: imageName } })
-        .run()
-    } catch (err) {
-      toast.error(extractIpcErrorMessage(err, 'Failed to insert image.'))
-    }
-  }, [editor, filePath])
+  const handleLocalImagePick = useLocalImagePick(editor, filePath)
 
   useEffect(() => {
     handleLocalImagePickRef.current = handleLocalImagePick
